Fix react query stuff
