+++ app/src/components/details/edit/DeleteJournalAction.tsx
@@ -3,19 +3,23 @@ import { IJournal } from "../../../serverApi/IJournal";
 import { Typography } from "@mui/material";
 import { useDeleteJournalMutation } from "../../../serverApi/reactQuery/mutations/useDeleteJournalMutation";
 import { DeleteButtons } from "../../common/DeleteButtons";
-import { queryKeysFactory } from "../../../serverApi/reactQuery/queryKeysFactory";
-import { useAppContext } from "../../../AppContext";
-import { useQueryClient } from "@tanstack/react-query";
 import { useItemAction } from "../../common/actions/searchParamHooks";
+import { useLocation, useNavigate } from "react-router-dom";
 
 export const DeleteJournalAction: React.FC<{
   journal: IJournal;
 }> = ({ journal }) => {
-  const { setAppAlert } = useAppContext();
-  const queryClient = useQueryClient();
   const { closeAction } = useItemAction();
+  const loc = useLocation();
+  const navigate = useNavigate();
 
-  const deleteJournalMutation = useDeleteJournalMutation(journal.id);
+  const deleteJournalMutation = useDeleteJournalMutation(journal.id, () => {
+    if (loc.pathname.startsWith(`/journals/details/${journal.id}`)) {
+      navigate("/");
+    } else {
+      closeAction();
+    }
+  });
 
   return (
     <>
@@ -27,25 +31,7 @@ export const DeleteJournalAction: React.FC<{
         entityType="journal"
         requiresConfirmation={true}
         onCancel={closeAction}
-        onDelete={() =>
-          deleteJournalMutation.mutate({
-            onSuccess: async () => {
-              closeAction();
-
-              setAppAlert({
-                title: `Successfully deleted journal.`,
-                type: "success",
-              });
-
-              await queryClient.invalidateQueries({
-                queryKey: queryKeysFactory.journal(journal.id),
-              });
-              await queryClient.invalidateQueries({
-                queryKey: queryKeysFactory.journals(),
-              });
-            },
-          })
-        }
+        onDelete={() => deleteJournalMutation.mutate()}
       />
     </>
   );
